fix NaN bug encountered when setting cutoff variance extremely high
When setting the cutoff variance to something absurd like 40, I
encountered the following error:
`Error in svd(x, nu = 0, nv = k) : infinite or missing values in 'x'`
Upon tracking down the problem, I saw that a dataframe had a column that
contained the same number. This means that when I calculate:
`(x - mean(x)) / (sqrt(sd(x)))`, I will get a `NaN` resulting in an
error in the svd. The fix was to check for `NaN`s before proceeding with
`svd` using a simple `is.na` function.